+++ tests/snapshots/arrays.rs
@@ -1,12 +1,12 @@
 // Arrays — fixed size, 2D, and safe access with .get()
 
 fn main() {
-    let mut scores: [i32; 5] = [0; 5];
+    let mut scores: [i64; 5] = [0; 5];
     scores[0] = 90;
     scores[1] = 85;
     scores[2] = 78;
     println!("{}", format!("{}{}", "scores[1] = ", scores[1]));
-    let mut grid: [[i32; 3]; 2] = [[0; 3]; 2];
+    let mut grid: [[i64; 3]; 2] = [[0; 3]; 2];
     grid[1][2] = 42;
     println!("{}", format!("{}{}", "grid[1][2] = ", grid[1][2]));
     // .get() returns an Option, so out-of-bounds is handled, not a crash
