+++ tests/snapshots/byref.rs
@@ -2,12 +2,12 @@
 // dereferences the parameter inside the function.
 
 fn main() {
-    let mut total: i32 = 0;
+    let mut total: i64 = 0;
     add_to(&mut total, 5);
     add_to(&mut total, 10);
     println!("{}", format!("{}{}", "total = ", total));
 }
 
-fn add_to(target: &mut i32, amount: i32) {
+fn add_to(target: &mut i64, amount: i64) {
     *target = *target + amount;
 }
