+++ tests/snapshots/methods.rs
@@ -2,7 +2,7 @@
 
 struct Person {
     pub name: String,
-    pub age: i32,
+    pub age: i64,
 }
 
 impl Person {
